+++ Ui_clip_handler.py
