+++ clip_handler.py
