+++ resources.qrc
@@ -1,0 +1,22 @@
+<RCC>
+  <qresource>
+    <file>icons/gauge.with.needle.fill.png</file>
+    <file>icons/filemenu.and.selection.png</file>
+    <file>icons/pencil.png</file>
+    <file>icons/chevron.compact.down.png</file>
+    <file>icons/chevron.compact.up.png</file>
+    <file>icons/app_icon.icns</file>
+    <file>icons/backward.fill.png</file>
+    <file>icons/custom.play.fill.png</file>
+    <file>icons/forward.fill.png</file>
+    <file>icons/chevron.down.png</file>
+    <file>icons/chevron.right.png</file>
+    <file>icons/arrowtriangle.down.fill.png</file>
+    <file>icons/pause.fill.png</file>
+    <file>icons/record.circle.action.png</file>
+    <file>icons/record.circle.png</file>
+    <file>icons/speaker.slash.fill.png</file>
+    <file>icons/speaker.wave.3.fill.on.png</file>
+    <file>icons/speaker.wave.3.fill.png</file>
+  </qresource>
+</RCC>
